+++ CHANGELOG.md
@@ -17,6 +17,8 @@ etc.
 - This should fix issues with images that don't properly load crashing the app.
 - Images that are loaded in comments show an activity indicator
 - Activity indicator for images in the feeds is animated properly
+- Better login validation
+- Fix issue where the wrong background color was displayed while opening the drawer or rotating the device
 
 ## [Version 1.1.1 (79)] - 2023-11-05T02:21:26Z